jwt-validation/requirements.txt
appdirs==1.4.4
black==21.7b0
blacksheep==1.0.9
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
essentials==1.1.4
essentials-openapi==0.1.5
guardpost==0.0.7
httptools==0.2.0
itsdangerous==1.1.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.8.3
rodi==1.1.1
six==1.16.0
tomli==1.2.1
